+++ src/ts/components/fullscreentogglebutton.ts
@@ -11,7 +11,7 @@ export class FullscreenToggleButton extends ToggleButton<ToggleButtonConfig> {
 
     this.config = this.mergeConfig(config, {
       cssClass: 'ui-fullscreentogglebutton',
-      text: 'Fullscreen'
+      text: 'Fullscreen',
     }, this.config);
   }