+++ src/ts/player.d.ts
@@ -23,8 +23,20 @@ declare namespace bitmovin {
         SINGLE: PlayerAPI.VR.ContentType,
         TAB: PlayerAPI.VR.ContentType,
         SBS: PlayerAPI.VR.ContentType,
-      }
-    }
+      },
+      STATE: {
+        READY: PlayerAPI.VR.State,
+        PLAYING: PlayerAPI.VR.State,
+        ERROR: PlayerAPI.VR.State,
+        UNINITIALIZED: PlayerAPI.VR.State,
+      },
+      TRANSITION_TIMING_TYPE: {
+        NONE: PlayerAPI.VR.TransitionTimingType,
+        EASE_IN: PlayerAPI.VR.TransitionTimingType,
+        EASE_OUT: PlayerAPI.VR.TransitionTimingType,
+        EASE_IN_OUT: PlayerAPI.VR.TransitionTimingType,
+      },
+    };
   }
 
   const player: PlayerStatic;
@@ -493,7 +505,19 @@ declare namespace bitmovin {
         TAB: PlayerAPI.VR.ContentType,
         SBS: PlayerAPI.VR.ContentType,
       },
-    }
+      STATE: {
+        READY: PlayerAPI.VR.State,
+        PLAYING: PlayerAPI.VR.State,
+        ERROR: PlayerAPI.VR.State,
+        UNINITIALIZED: PlayerAPI.VR.State,
+      },
+      TRANSITION_TIMING_TYPE: {
+        NONE: PlayerAPI.VR.TransitionTimingType,
+        EASE_IN: PlayerAPI.VR.TransitionTimingType,
+        EASE_OUT: PlayerAPI.VR.TransitionTimingType,
+        EASE_IN_OUT: PlayerAPI.VR.TransitionTimingType,
+      },
+    };
 
     vr: PlayerAPI.PlayerVRAPI;
   }
@@ -913,7 +937,14 @@ declare namespace bitmovin {
           /**
            *  Two equirectangular videos for 3D content in side-by-side position.
            */
-        SBS
+        SBS,
+      }
+
+      enum State {
+        READY,
+        PLAYING,
+        ERROR,
+        UNINITIALIZED,
       }
 
       /**